Refactor: miscellaneous
Some enhancements added while importing bergamot propogates into source.
Driven by https://github.com/jerinphilip/slimt/pull/7.
