+++ slimt/Aligned.cc
@@ -10,8 +10,12 @@ Aligned::Aligned(size_t alignment, size_t size)
 Aligned::~Aligned() { release(); }
 
 void* Aligned::data() const { return data_; }
-
 size_t Aligned::size() const { return size_; }
+
+char* Aligned::begin() const { return reinterpret_cast<char*>(data_); }
+char* Aligned::end() const { return reinterpret_cast<char*>(data_) + size_; }
+
+bool Aligned::empty() const { return size_ == 0; }
 
 Aligned::Aligned(Aligned&& from) noexcept {
   if (this != &from) {
