+++ slimt/Shortlist.hh
@@ -3,20 +3,18 @@
 
 #include "slimt/Batch.hh"
 #include "slimt/Io.hh"
+#include "slimt/Types.hh"
 #include "slimt/Vocabulary.hh"
 
 namespace slimt {
 
 class Shortlist {
  public:
-  using Word = Vocabulary::Word;
-  using Words = Vocabulary::Words;
-
   explicit Shortlist(Words words) : words_(std::move(words)) {}
-  const std::vector<Vocabulary::Word>& words() const { return words_; }
-  Vocabulary::Word reverse_map(int idx) { return words_[idx]; }
+  const std::vector<Word>& words() const { return words_; }
+  Word reverse_map(int idx) { return words_[idx]; }
 
-  int try_forward_map(Vocabulary::Word w_idx) {
+  int try_forward_map(Word w_idx) {
     auto first = std::lower_bound(words_.begin(), words_.end(), w_idx);
     if (first != words_.end() && *first == w_idx) {
       // Check if element not less than w_idx has been found
@@ -36,9 +34,9 @@ class Shortlist {
 
 class ShortlistGenerator {
  public:
-  using Word = Shortlist::Word;
-  using Words = Shortlist::Words;
-  static const uint64_t kMagic = 0xF11A48D5013417F5;
+  static constexpr uint64_t kMagic = 0xF11A48D5013417F5;
+  static constexpr uint64_t kFrequent = 100;
+  static constexpr uint64_t kBest = 100;
 
   // construct directly from buffer
   ShortlistGenerator(
@@ -57,8 +55,8 @@ class ShortlistGenerator {
   size_t source_index_;
   bool shared_{false};
 
-  uint64_t first_num_{100};  // baked into binary header
-  uint64_t best_num_{100};   // baked into binary header
+  uint64_t frequent_{kFrequent};  // baked into binary header
+  uint64_t best_{kBest};          // baked into binary header
 
   // shortlist is stored in a skip list
   // [&shortLists_[word_to_offset_[word]],
@@ -70,13 +68,13 @@ class ShortlistGenerator {
   uint64_t shortlist_size_;
 
   const uint64_t* word_to_offset_;
-  const Vocabulary::Word* shortlist_;
+  const Word* shortlist_;
 
   struct Header {
     uint64_t magic;                // BINARY_SHORTLIST_MAGIC
-    uint64_t checksum;             // hash([&first_num, eof]).
-    uint64_t first_num;            // Limits used to create the shortlist.
-    uint64_t best_num;             //
+    uint64_t checksum;             // hash([&frequent, eof]).
+    uint64_t frequent;             // Limits used to create the shortlist.
+    uint64_t best;                 //
     uint64_t word_to_offset_size;  // Length of word_to_offset_ array.
     uint64_t shortlist_size;       // Length of short_lists_ array.
   };
