+++ slimt/Vocabulary.cc
@@ -14,8 +14,8 @@ Vocabulary::Vocabulary(const std::string &fpath) {
   processor_.Load(fpath);
 }
 
-std::tuple<Vocabulary::Words, Vocabulary::Views> Vocabulary::encode(
-    const std::string_view &line, bool add_eos /* = false*/) {
+std::tuple<Words, Views> Vocabulary::encode(const std::string_view &line,
+                                            bool add_eos /* = false*/) const {
   absl::string_view a_line(line.data(), line.size());
   std::vector<absl::string_view> views;
   sentencepiece::SentencePieceText sentencepiece_text;
@@ -57,8 +57,8 @@ std::tuple<Vocabulary::Words, Vocabulary::Views> Vocabulary::encode(
   return {std::move(words), std::move(std_views)};
 }
 
-std::tuple<std::string, Vocabulary::Views> Vocabulary::decode(
-    const Words &words, bool ignore_eos) {
+std::tuple<std::string, Views> Vocabulary::decode(const Words &words,
+                                                  bool ignore_eos) const {
   sentencepiece::SentencePieceText sentencepiece_text;
   std::vector<std::string_view> views;
