+++ slimt/slimt.hh
@@ -1,6 +1,7 @@
 #pragma once
 #include "slimt/Batch.hh"
 #include "slimt/Io.hh"
+#include "slimt/Macros.hh"
 #include "slimt/Model.hh"
 #include "slimt/QMM.hh"
 #include "slimt/Shortlist.hh"
